+++ models/CNNModelTest56.cs
@@ -7,7 +7,7 @@ namespace dl
 {
     class CNNModelTest56 : IModel
     {
-        public Func<IEnumerable<Tuple<double, double>>, double> ErrorFunction { get; set; }
+        public Func<IEnumerable<Tuple<float, float>>, float> ErrorFunction { get; set; }
 
         public IEnumerable<ILayer> Layers { get; set; }
